requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.9
pynput>=1.7.6
pyobjc-framework-Quartz>=10.0
pyobjc-framework-Cocoa>=10.0
numpy>=1.24.0
